完善代码，修改bug

+++ src/main/java/com/sc/qisi_system/module/admin/entity/SysRoleMenu.java
@@ -32,7 +32,7 @@ public class SysRoleMenu implements Serializable {
     /**
      * 业务身份
      */
-    private Long identityId;
+    private Integer identityId;
 
 
     /**
